requirements.txt
discord.py
python-dotenv      
aiohttp             
asyncio              
colorama            
requests    